requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.11
redis==5.3.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
alembic==1.17.2
sqlalchemy==2.0.36
